funcCallCount

+++ unidock/unidock_engine/src/cuda/common.cu
@@ -3,7 +3,7 @@
 //
 #include "common.cuh"
 
-__device__ __managed__ unsigned int funcCallCount = 0;
+__device__ __managed__ unsigned long long funcCallCount = 0ULL;
 
 
 __constant__ bool FLAG_CONSTRAINT_DOCK = false;
